refactor: combobox
