--- src/components/Combobox-compound/index.ts
@@ -1,1 +1,0 @@
-export * from './combobox';
