+++ src/components/Combobox/index.ts
@@ -1,0 +1,10 @@
+import { Combobox, ComboboxOption, ComboboxOptions, ComboboxTrigger } from './combobox';
+
+const ComboboxCompound = Object.assign(Combobox, {
+  //   label: ComboboxLabel,
+  Trigger: ComboboxTrigger,
+  Options: ComboboxOptions,
+  Option: ComboboxOption,
+});
+
+export { ComboboxCompound as Combobox };
